combined spec helpers
