+++ spec/requests/bridges_request_spec.rb
@@ -1,12 +1,10 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
-require './spec/support/users_helper'
-require './spec/support/bridges_helper'
+require './spec/support/main_helper'
 
 RSpec.configure do |c|
-  c.include UsersHelper
-  c.include BridgesHelper
+  c.include MainHelper
 end
 
 RSpec.describe 'Bridges', type: :request do
